refactory: modifying repos github request location
